# swagger.json
{"openapi":"3.0.3","info":{"title":"Transcendence API","description":"API documentation","version":"1.0.0"},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{}},"paths":{"/user/{id}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/user/leaderboard":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/user/me":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/user/{id}/gamehistory":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/user/register":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userName":{"type":"string"},"displayName":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}},"required":["userName","displayName","password"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/user/enable2fa":{"put":{"responses":{"200":{"description":"Default Response"}}}},"/user/verify2fa":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/user/login":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userName":{"type":"string"},"password":{"type":"string"}},"required":["userName","password"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/user/login2fa":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userName":{"type":"string"},"otp":{"type":"string"}},"required":["userName","otp"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/user/updateDisplayName":{"put":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"displayName":{"type":"string"}},"required":["displayName"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/user/updatePassword":{"put":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"password":{"type":"string"},"newPassword":{"type":"string"}},"required":["password, newPassword"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/user/uploadpicture":{"put":{"description":"Route unusable from Swagger, test with postman / curl","responses":{"200":{"description":"Default Response"}}}},"/game/":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"player1Id":{"type":"number"},"player2Id":{"type":"number"}},"required":["player1Id","player2Id"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/game/start":{"patch":{"responses":{"200":{"description":"Default Response"}}}},"/game/finish":{"patch":{"responses":{"200":{"description":"Default Response"}}}},"/game/{id}/match":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/game/pong/generateGameId":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/chat/dm/{peerid}/messages":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"peerid","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/dm/{peerid}/block":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"peerid","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chatdm/{peer}/unblock":{"delete":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"peer","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/dm/{peerid}/read":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"peerid","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/unread":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/list":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/{friendId}/sendRequest":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"friendId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/{friendId}/acceptRequest":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"friendId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/{friendId}/rejectRequest":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"friendId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/{friendId}/removeFriend":{"delete":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"friendId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/chat/friends/status":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/tournament/list":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/status":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/participants":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/matches":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{tournamentId}/match/{matchId}/result":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"tournamentId","required":true},{"schema":{"type":"string"},"in":"path","name":"matchId","required":true}],"responses":{"200":{"description":"Default Response"}}},"put":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"winnerId":{"type":"number"},"player1Score":{"type":"number"},"player2Score":{"type":"number"}},"required":["winnerId","player1Score","player2Score"]}}},"required":true},"parameters":[{"schema":{"type":"string"},"in":"path","name":"tournamentId","required":true},{"schema":{"type":"string"},"in":"path","name":"matchId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{tournamentId}/match/{matchId}/gameId":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"tournamentId","required":true},{"schema":{"type":"string"},"in":"path","name":"matchId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/create":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/join":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"displayname":{"type":"string"}},"required":["displayname"]}}},"required":true},"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/leave":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"number"},"displayname":{"type":"string"}},"anyOf":[{"required":["userId"]},{"required":["displayname"]}]}}}},"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/start":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/tournament/{id}/finish":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}}},"security":[{"bearerAuth":[]}]}